move quota to dedicated service
